+++ NEWS.md
@@ -1,3 +1,5 @@
+# Unreleased Changes
+
 # News in version 0.4.0
 
 - Prepare `DBFixture` for SQLAlchemy 2.0.