+++ NEWS.md
@@ -1,4 +1,4 @@
-# Unreleased Changes
+# News in version 0.4.1
 
 - Fix various `DBFixture` methods.
 - Fix various SQLAlchemy 2.0 compatibility issues in `DBFixture`.